@bsalamat 2018-08-08
I tried running our AntiAffinity scheduler benchmark:
```
make test-integration WHAT=./test/integration/scheduler_perf GOFLAGS="-v=1" KUBE_TEST_VMODULE="''" KUBE_TEST_ARGS="-run=xxx -bench BenchmarkSchedulingAntiAffinity/5000Nodes/0Pods"
```

and it seems that this PR in its current form does not improve performance when scheduling 2000 pods with AntiAffinity in a cluster with 5000 nodes. Here are the results when eCache is disabled:

With this PR:
```
goos: linux
goarch: amd64
pkg: k8s.io/kubernetes/test/integration/scheduler_perf
BenchmarkSchedulingAntiAffinity/5000Nodes/0Pods-12         	    2000	 299407004 ns/op
PASS
ok  	k8s.io/kubernetes/test/integration/scheduler_perf	612.984s
```

Without this PR:
```
goos: linux
goarch: amd64
pkg: k8s.io/kubernetes/test/integration/scheduler_perf
BenchmarkSchedulingAntiAffinity/5000Nodes/0Pods-12         	    2000	 298837178 ns/op
PASS
ok  	k8s.io/kubernetes/test/integration/scheduler_perf	611.690s
```