@mohamed-mehany 2018-08-09
/retest pull-kubernetes-e2e-kops-aws