@bsalamat 2018-08-11
I found sometime to profile the scheduler with and without the changes in this PR. When we run the BenchmarkSchedulingAntiAffinity without this PR, about 8% of the time is spent in `satisfiesExistingPodsAntiAffinity` and almost the rest is spent in `satisfiesPodsAffinityAntiAffinity` function. The latter is the function that checks the anti-affinity rules of the pod being scheduled. This PR optimizes `satisfiesExistingPodsAntiAffinity` which is a much smaller part of this benchmark. With this PR, `satisfiesExistingPodsAntiAffinity` has disappeared from the graphical output of pprof, which means that the function no longer takes a significant amount of time. So, this PR has worked and has optimized the function that it is targeting, but given that the function is not a large part of this particular benchmark, we don't see much performance improvement.
If we apply a similar optimization to checking anti-affinity rules of the pod being scheduled, we will probably see much better results.