@mohamed-mehany 2018-08-13
/test pull-kubernetes-integration