@mohamed-mehany 2018-08-14
/test pull-kubernetes-e2e-gce-device-plugin-gpu